requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
pyTelegramBotAPI==4.24.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
Telethon==1.37.0
urllib3==2.2.3